Move timestamp module to folder
