+++ lib/lotus/model.rb
@@ -2,7 +2,7 @@ require 'lotus/model/version'
 require 'lotus/entity'
 require 'lotus/entity/dirty_tracking'
 require 'lotus/repository'
-require 'lotus/repository_timestamps'
+require 'lotus/repository/timestamps'
 require 'lotus/model/mapper'
 require 'lotus/model/configuration'
